+++ src/app/api/_utils/jwt/jwtPayloadType.ts
@@ -1,4 +1,4 @@
-export type jwtPayload = {
+export type jwtPayloadType = {
   handle: string;
   server: string;
   jwtIndex: number;
