+++ src/app/api/db/create-question/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from 'next/server';
-import { CreateQuestionApiService } from '@/api/db/create-question/_service';
+import { CreateQuestionApiService } from '@/_service/question/create-question-service';
 
 export async function POST(req: NextRequest) {
   const service = CreateQuestionApiService.get();
