+++ src/app/api/db/delete-answer/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from 'next/server';
-import { AnswerService } from '../_answer-service/_answer-service';
+import { AnswerService } from '@/_service/answer/answer-service';
 
 export async function POST(req: NextRequest) {
   const service = AnswerService.get();
