+++ src/app/api/db/fetch-my-profile/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from 'next/server';
-import { ProfileService } from '@/app/api/db/_profile-service/profile-service';
+import { ProfileService } from '@/app/api/_service/profile/profile-service';
 
 export async function GET(req: NextRequest) {
   const service = ProfileService.get();
