+++ src/app/api/user/blocking/create/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from 'next/server';
-import { BlockingService } from '@/api/user/blocking/_service';
+import { BlockingService } from '@/app/api/_service/blocking/blocking-service';
 
 export async function POST(req: NextRequest) {
   const service = BlockingService.get();
