+++ src/app/api/user/following/list/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from 'next/server';
-import { FollowingService } from '../_service';
+import { FollowingService } from '../../../_service/following/following-service';
 
 const service = FollowingService.get();
 export async function POST(req: NextRequest) {
