+++ src/app/api/web/misskey-login/route.ts
@@ -5,7 +5,7 @@ import { MiApiError, MiAuthSession } from '@/app';
 import detectInstance from '@/utils/detectInstance/detectInstance';
 import { GetPrismaClient } from '@/app/api/_utils/getPrismaClient/get-prisma-client';
 import { Logger } from '@/utils/logger/Logger';
-import { RateLimiterService } from '@/app/api/_utils/ratelimiter/rateLimiter';
+import { RateLimiterService } from '@/_service/ratelimiter/rateLimiterService';
 import { getIpHash } from '@/app/api/_utils/getIp/get-ip-hash';
 import { getIpFromRequest } from '@/app/api/_utils/getIp/get-ip-from-Request';
 import { sendApiError } from '@/app/api/_utils/apiErrorResponse/sendApiError';
