+++ src/instrumentation.ts
@@ -1,6 +1,6 @@
 export async function register() {
   if (process.env.NODE_ENV === 'production') {
-    const qs = await import('@/api/_queue-service/queueService');
+    const qs = await import('@/_service/queue/queueService');
     qs.QueueService.get();
 
     const env_arr = ['REDIS_HOST', 'REDIS_PORT', 'WEB_URL', 'DATABASE_URL', 'JWT_SECRET', 'NOTI_TOKEN', 'NOTI_HOST'];
